Moved helpers to core, added license headers
